Fixes to acceleration\velocity handling, fixed main loop tick miscalculation

+++ Moottori/include/Component/LocationComponent.h
@@ -5,7 +5,7 @@ class LocationComponent : public Component
 {
 public:
 	LocationComponent() : mX(0), mY(0), mRotation(0) {} 
-	LocationComponent(int x, int y) : mX(x), mY(y), mRotation(0) {}
+	LocationComponent(double x, double y) : mX(x), mY(y), mRotation(0) {}
 	virtual ~LocationComponent() {}
 
 #if !defined _MSC_VER || _MSC_VER >= 1800 
@@ -18,11 +18,11 @@ private:
 public:
 #endif
 
-	int GetX() const { return mX; }
-	void SetX(int x) { mX = x; }
+	int GetX() const { return (int)mX; }
+	void SetX(double x) { mX = x; }
 
-	int GetY() const { return mY; }
-	void SetY(int y) { mY = y; }
+	int GetY() const { return (int)mY; }
+	void SetY(double y) { mY = y; }
 
 	double GetRotation() { return mRotation; }
 	void SetRotation(double r) { mRotation = r; }
@@ -31,7 +31,7 @@ protected:
 	void OnEventHandlerRegistration() override;
 private:
 	void HandleLocationChangeEvents(Event *event);
-	int mX;
-	int mY;
+	double mX;
+	double mY;
 	double mRotation;
 };
